3rdparty: Add the 'win32' platform

+++ 3rdparty.sh
@@ -35,6 +35,13 @@ Qt4_archive="$Qt4_name.tar.gz"
 Qt4_sources="http://download.qt.io/official_releases/qt/4.8/4.8.7/$Qt4_archive"
 
 #--------------------------------------------------------------------------------------------------
+# Windows
+
+win32_archive="3rdparty.tar.gz"
+
+win32_sources="http://omega.gg/get/Sky/3rdparty/win32"
+
+#--------------------------------------------------------------------------------------------------
 # Linux
 
 usr="/usr/lib"
@@ -81,9 +88,9 @@ if [ $# != 2 ] || [ $1 != "all"       -a \
                     $1 != "install"   -a \
                     $1 != "uninstall" -a \
                     $1 != "deploy"    -a \
-                    $1 != "clean" ] || [ $2 != "ubuntu" ]; then
+                    $1 != "clean" ] || [ $2 != "win32" -a $2 != "ubuntu" ]; then
 
-    echo "Usage: 3rdparty <all | install | uninstall | deploy | clean> <ubuntu>"
+    echo "Usage: 3rdparty <all | install | uninstall | deploy | clean> <win32 | ubuntu>"
 
     exit 1
 fi
@@ -206,12 +213,15 @@ fi
 
 if [ $1 = "clean" ]; then
 
-    echo "CLEANING"
+    if [ $linux = true ]; then
 
-    rm -rf "$Qt4"
-    rm -rf "$VLC"
-    rm -rf "$libtorrent"
-    rm -rf "$Boost_linux"
+        echo "CLEANING"
+
+        rm -rf "$Qt4"
+        rm -rf "$VLC"
+        rm -rf "$libtorrent"
+        rm -rf "$Boost_linux"
+    fi
 
     exit 0
 fi
@@ -222,7 +232,19 @@ fi
 
 if [ $1 = "all" ] || [ $1 = "deploy" ]; then
 
-    if [ $linux = true ]; then
+    if [ $2 = "win32" ]; then
+
+        echo "DEPLOYING 3rdparty"
+
+        cd ..
+
+        curl -o "$win32_archive" "$win32_source"
+
+        tar -xvzf "$win32_archive"
+
+        rm "$win32_archive"
+
+    elif [ $linux = true ]; then
 
         echo "DEPLOYING Qt"
